Add dev notary key for testing against devnet (#1)
Added dev key to be used in Solana Dev network for a second network in
Corda Solana program. The aim is to partition Cordapps using tokens SDK
from others using dev key from Corda Enterprise.
Instructions how to consume this key in projects with Cordapps (for
running via deployNodes etc) will be added in other PRs.
The key in not yet whitelisted in Corda program on Solana Devnet.