Reducing configuration of Kotlin modules (#10)
Introduced a `default-kotlin` buildSrc plugin which has the boilerplate
for a Kotlin module. This includes JUnit (and a nifty test logger) and
Detekt plugin. Not everything could be moved across due to limitations
with the Gradle version we are using.

I've also pull across the Deteket config that's ended up in
https://github.com/corda/solana-notary/. Hopefully this will have some
sensible rules going forward.