Fix broken CI (#29)
A couple of issues with the build:

* Sometimes a solana-test-validator is left running from a previous run,
probably from a failed test, which causes issues on the next run. So now
we first force kill any leftover instances.
* Because we're using SNAPSHOT of Corda Enterprise, if a newer SNAPSHOT
build is published to Artifactory which makes some internal breaking
changes, then that conflicts with the explicit SNAPSHOT build we're
using for the node-driver. So now we add the ENT node as a
`testRuntimeOnly` dependency to keep the transitive node dependency
in-sync.
* Build logs were showing `Expiring Daemon due to JVM Metaspace space
being exhausted` messages. Doesn't seem to be causing an issues, but
maybe reduced build performance. Increased metaspace size of the deamon.

Also, added either `os` or `ent` to the various Corda libs we're using
to make it clear which is one is being used.