ENT-14629 - Supporting multiple token states in redemption (#41)
This PR implements the support for multiple fungible tokens in
redemption. So far the assumption was that only one fungible token will
satisfy the redeemed amount. That is not the case in real life as the
token selection process from the Token SDK can select multiple tokens
from the vault.