ENT-14611  Bridging recovery - preventing missed updates during node startup (#42)
1. Start vault listener for new tokens to bridge before starting a
recovery on node startup to avoid missing updates.
2. Wrap Bridge flow in `try catch` and send to it to a flow hospital for
any error that is not `Notary.Conflict`.
This change is in relation to 1st one, we may start a flow for the
same`FungibleToken` twice - 1st from a vault update and the 2nd time
when it's collected by a startup query and considered as "missed" one -
this may happen when a new update arrives just during a node startup.
This will cause an issue with moving a token to Confidential Identity
manifesting as `Notary.Conflict` error - which is fine.

This was the only issue identified with bridging recovery.

This was manually tested by e.g adding duplicates code to move to CI and
forcing `Notary.Conflict` error:
```
  val moveTx2 = subFlow(
                MoveAndLockFungibleTokenFlow(
                    token,
                    bridgingCoordinates,
                    lockingHolder,
                    participantSessions = sessionsForParties(listOf(lockingHolder)),
                    observerSessions = sessionsForParties(observers),
                )
            )
```