ENT-14531 Redemption Recovery (#38)
This PR introduces a recovery mechanism implementation for the
redemption process. Mainly there are two main steps that are taken into
account with this implementation.

1. At start-up (and then periodically every
`redemptionCheckIntervalSeconds`) the `BridgingService` scans all Solana
redemption account balances and if any of those is positive it will
trigger the `RedeemFungibleTokenFlow` execution - This addresses the
case when node was down for a period of time in which the Solana burn
accounts were modified or if we missed any websocket notifications
2. At the end of the `RedeemFungibleTokenFlow` execution, we check if
there is a positive balance on the processed redemption account - This
addresses the case when the node stopped in the middle of the
`RedeemFungibleTokenFlow` execution