Refactor some methods that will be reused by Driver tests. (#47)
Methods were take outside the class as extension methods of
`SolanatTestValidator` class. These could be moved to new file as well.
Optional second change:
This causes a method call must be preceded by an object name, and it
makes a line longer, so I renamed `testValidator` to a `validator`. I
think it's fine, from test business logic it only important to denote
that we are using to a validator, and actually it gives more info to
reader that a method deals with validator, e.g:
`transfer(stakeholderInfo.signer, fromTokenAccount, ....`
is less informative than:
`validator.transfer(stakeholderInfo.signer, fromTokenAccount, ...`