ENT-14562 Publish to Artifactory (#7)
Context: this will publish 2 Cordapps - contracts and workflows for
imported by any Corda Node which needs connect to Solana. The publishing
process uses the axion plugin:
https://axion-release-plugin.readthedocs.io/en/latest/configuration/publishing/

I aim to publish JARs on each merged commit to `main` - it probably
different than usually we do nightly builds.
Publishing destination: `r3-corda-dev` for SNAPSHOTs and
`r3-corda-releases` for RELEASEs

Publishing to `r3-corda-dev works:
https://software.r3.com/ui/native/r3-corda-dev/com/r3/corda/lib/solana/
It is triggered on the main merge. 

It will be also possible to run the publishing pipeline once we merge
the changes to the main branch and the dedicated Jenkins configuration
is created.

---------

Co-authored-by: Michal Kit <michal.kit@r3.com>